requirement.txt
cycler==0.10.0
dlib==19.18.0
joblib==0.14.0
matplotlib==1.5.1
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.1.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
sklearn==0.0
torch==1.2.0
torchvision==0.4.0
